change float x int

+++ add_two.py
@@ -1,4 +1,4 @@
 # Program that asks for a number and adds 2 to it
-number = float(input("Enter a number: "))
+number = int(input("Enter a number: "))
 result = number + 2
 print(f"The result is: {result}")
